# infra/database/query/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package query

import (
	"time"
)

type Comment struct {
	ID          int64
	PostID      int64
	Body        string
	CommentedAt time.Time
}

type Post struct {
	ID       int64
	Title    string
	Content  string
	PostedAt time.Time
}