# infra/handler/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"fmt"
	"time"
)

func (s *ErrorStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// Ref: #/components/schemas/Comment
type Comment struct {
	ID          int64     `json:"id"`
	PostId      int64     `json:"postId"`
	Body        string    `json:"body"`
	CommentedAt time.Time `json:"commentedAt"`
}

// GetID returns the value of ID.
func (s *Comment) GetID() int64 {
	return s.ID
}

// GetPostId returns the value of PostId.
func (s *Comment) GetPostId() int64 {
	return s.PostId
}

// GetBody returns the value of Body.
func (s *Comment) GetBody() string {
	return s.Body
}

// GetCommentedAt returns the value of CommentedAt.
func (s *Comment) GetCommentedAt() time.Time {
	return s.CommentedAt
}

// SetID sets the value of ID.
func (s *Comment) SetID(val int64) {
	s.ID = val
}

// SetPostId sets the value of PostId.
func (s *Comment) SetPostId(val int64) {
	s.PostId = val
}

// SetBody sets the value of Body.
func (s *Comment) SetBody(val string) {
	s.Body = val
}

// SetCommentedAt sets the value of CommentedAt.
func (s *Comment) SetCommentedAt(val time.Time) {
	s.CommentedAt = val
}

// Ref: #/components/schemas/CommentList
type CommentList struct {
	Items []Comment `json:"items"`
}

// GetItems returns the value of Items.
func (s *CommentList) GetItems() []Comment {
	return s.Items
}

// SetItems sets the value of Items.
func (s *CommentList) SetItems(val []Comment) {
	s.Items = val
}

// CommentsDeleteNoContent is response for CommentsDelete operation.
type CommentsDeleteNoContent struct{}

// Ref: #/components/schemas/CreateCommentRequest
type CreateCommentRequest struct {
	Body string `json:"body"`
}

// GetBody returns the value of Body.
func (s *CreateCommentRequest) GetBody() string {
	return s.Body
}

// SetBody sets the value of Body.
func (s *CreateCommentRequest) SetBody(val string) {
	s.Body = val
}

// Ref: #/components/schemas/CreatePostRequest
type CreatePostRequest struct {
	Title   string `json:"title"`
	Content string `json:"content"`
}

// GetTitle returns the value of Title.
func (s *CreatePostRequest) GetTitle() string {
	return s.Title
}

// GetContent returns the value of Content.
func (s *CreatePostRequest) GetContent() string {
	return s.Content
}

// SetTitle sets the value of Title.
func (s *CreatePostRequest) SetTitle(val string) {
	s.Title = val
}

// SetContent sets the value of Content.
func (s *CreatePostRequest) SetContent(val string) {
	s.Content = val
}

// Ref: #/components/schemas/Error
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// GetCode returns the value of Code.
func (s *Error) GetCode() int32 {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *Error) GetMessage() string {
	return s.Message
}

// SetCode sets the value of Code.
func (s *Error) SetCode(val int32) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *Error) SetMessage(val string) {
	s.Message = val
}

// ErrorStatusCode wraps Error with StatusCode.
type ErrorStatusCode struct {
	StatusCode int
	Response   Error
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorStatusCode) GetResponse() Error {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorStatusCode) SetResponse(val Error) {
	s.Response = val
}

// Ref: #/components/schemas/Post
type Post struct {
	ID       int64     `json:"id"`
	Title    string    `json:"title"`
	Content  string    `json:"content"`
	PostedAt time.Time `json:"postedAt"`
}

// GetID returns the value of ID.
func (s *Post) GetID() int64 {
	return s.ID
}

// GetTitle returns the value of Title.
func (s *Post) GetTitle() string {
	return s.Title
}

// GetContent returns the value of Content.
func (s *Post) GetContent() string {
	return s.Content
}

// GetPostedAt returns the value of PostedAt.
func (s *Post) GetPostedAt() time.Time {
	return s.PostedAt
}

// SetID sets the value of ID.
func (s *Post) SetID(val int64) {
	s.ID = val
}

// SetTitle sets the value of Title.
func (s *Post) SetTitle(val string) {
	s.Title = val
}

// SetContent sets the value of Content.
func (s *Post) SetContent(val string) {
	s.Content = val
}

// SetPostedAt sets the value of PostedAt.
func (s *Post) SetPostedAt(val time.Time) {
	s.PostedAt = val
}

// Ref: #/components/schemas/PostList
type PostList struct {
	Items []Post `json:"items"`
}

// GetItems returns the value of Items.
func (s *PostList) GetItems() []Post {
	return s.Items
}

// SetItems sets the value of Items.
func (s *PostList) SetItems(val []Post) {
	s.Items = val
}

// PostsDeleteNoContent is response for PostsDelete operation.
type PostsDeleteNoContent struct{}

// Ref: #/components/schemas/UpdateCommentRequest
type UpdateCommentRequest struct {
	Body string `json:"body"`
}

// GetBody returns the value of Body.
func (s *UpdateCommentRequest) GetBody() string {
	return s.Body
}

// SetBody sets the value of Body.
func (s *UpdateCommentRequest) SetBody(val string) {
	s.Body = val
}

// Ref: #/components/schemas/UpdatePostRequest
type UpdatePostRequest struct {
	Title   string `json:"title"`
	Content string `json:"content"`
}

// GetTitle returns the value of Title.
func (s *UpdatePostRequest) GetTitle() string {
	return s.Title
}

// GetContent returns the value of Content.
func (s *UpdatePostRequest) GetContent() string {
	return s.Content
}

// SetTitle sets the value of Title.
func (s *UpdatePostRequest) SetTitle(val string) {
	s.Title = val
}

// SetContent sets the value of Content.
func (s *UpdatePostRequest) SetContent(val string) {
	s.Content = val
}